adding in new powershell module and a new solution for merging two binary trees
