+++ CSharp/LeetCodeStuff/LeetCodeStuffTests/LeetCodeStuffTests.csproj
@@ -56,6 +56,7 @@
     <Compile Include="JewelsAndStonesTests.cs" />
     <Compile Include="JudgeCircleTests.cs" />
     <Compile Include="MaxBinaryTreeTests.cs" />
+    <Compile Include="MergeTwoBinaryTreesTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
